A c++ interface for the fortran routines
The interface is two layered: there's ABL.cpp, which are the c++
routines, as well as an "external C" declaration of the Fortran 2008
routines. Then there's neXtSIM_f2008_interface.f90, which is the Fortran
2008 interface to the FORTRAN 77 code. Reason being, that it was only in
the 2008 version of the Fortran standard that we got a clearly defined
interface between Fortran and C.

We may want to reconsider this, and just rewrite
Initialization_module_NeXtSIM.for and Command_module_for_NeXtSIM.for in
Fortran 2008.