ABL model as a stand-alone model
I'm experimenting with the idea of treating the ABL model as a
stand-alone model that talks to neXtSIM thorough OASIS. This has several
benefits (no moving mesh, simpler testing, helps with coupling to a full
atmosphere later ... and preserves the original Fortran code - at least
to begin with).

This commit adds ABL.for, which is nearly identical to the old
Combined.for (save one bug fix, it seems). Also adds sourceDiffuse.f -
which is needed for ABL.for to compile.