First step towards a 3D model
Here I've added one dimension to all the prognostic variables. It
remains to be seen what I'll do with 'theta' though, which is a
diagnostic, but an important one.

The idea is to preprocess all inputs so that they are interpolated on a
grid with all the land points removed. It would therefore be a vector
for surface variables and a 2D array for 3D variables. The model would
read this, along with placement information for all points so that it
can output on the original 2D grid. This allows me to write the
interpolation routines as pre-processing in python.