Change from vector to array for grid representation
Adhering to the KISS principle I'm keeping everything as an array,
instead of a vector. Reason being that I want to read in the forcing as
array and write out the results as array. I may also want to calculate
geostrophic velocities based on SLP, and then I need to keep the array
notation around.

Switching to vector may be faster in the end, but for now it's KISS (and
"premature optimisation is the root of all evil).