Routines for reading ERA5
This adds routines to read and interpolate ERA5 data. I'm using Fortran
types and type bound procedures to crete objects (of sorts). Each
variable is an object with two methods, init and read_input. The
interpolation is just a bilinear one on the lat/lon grid - which is good
enough for government busness.

TODO:
 - Calculate weights only once.
 - Create a type for reading in 3D data