Use a format string for input file name
This commit allows us to use a format string to give the name of the
(ERA5) forcing file to be read. It uses the string ${var} as a place
holder for the variable name and C/C++ strftime formats for the time
(%Y, %M, etc). Example "data/ERA5_${var}_y%Y.nc"