Lookup tables replaced by an generated trivial hash table
The slowest part of ABL.for was the two lookup tables in subfec and
subfec2. Here I replace those functions with a trivial-hash-table
implementation that takes advantage of the fact that the x value is
defined with a precision of 0.1. Using flooor and some simple integer
math I convert x into an array index and populate the array through
automatic code generation. There's also a test (run when you run make
test) that compares the outputs from the new and old implementations.
Speedup of about 2.5 of a single model time step.